ipod_wrapped/requirements.txt
polars
python-dotenv
requests